Fixed: java.lang.Long cannot be cast to java.lang.Integer

+++ app/models/Batch.java
@@ -113,7 +113,7 @@ public class Batch extends JpaModel {
 			int size = Play.application().configuration().getInt("page.size");
 			if (index < 1)
 				index = 1;
-			Integer rows = (Integer) JPA.em()
+			Long rows = (Long) JPA.em()
 					.createQuery("select count(*) from Batch")
 					.getSingleResult();
 			List<Batch> list = JPA.em()
