+++ app/models/Ingredient.java
@@ -79,7 +79,7 @@ public class Ingredient extends JpaModel {
 			int size = Play.application().configuration().getInt("page.size");
 			if (index < 1)
 				index = 1;
-			Integer rows = (Integer) JPA.em()
+			Long rows = (Long) JPA.em()
 					.createQuery("select count(*) from Ingredient")
 					.getSingleResult();
 			List<Ingredient> list = JPA.em()
